Add files via upload
For the analysis of Task 1 , there is a brief about what all methodologies and approaches I followed to derive insights and few recommendations for commercial and supply chain team.

For Task 2 , I have done a seasonality analysis of sales for countries and forecasted sales for a month. 
Have also presented few insights

Please refer attached python notebooks for detailed code.